Moved component to UI folder
