+++ src/components/UI/index.ts
@@ -1,0 +1,1 @@
+export { default as IconButton } from './IconButton';
