+++ src/components/index.ts
@@ -1,3 +1,3 @@
 export { default as AudioPlayer	} from './AudioPlayer';
 export { default as Header			} from './Header';
-export { default as IconButton	} from './IconButton';
+export { default as IconButton	} from './UI/IconButton';
